Fix error with device image on Firefox browser
When using a device image on Firefox the image would rarely appear as
the puzzle image even though the small puzzle preview image did display.

Found a parameter error when calling the toDataURL method on the final
canvas, Chrome and Safari seemed to ignore the issue and just display
the puzzle, whereas Firefox showed a blank puzzle.